dev I/O model first commit
Each device is declared with:
* an major/minor pair (8 bits each), coupled together as 16 bit "id";
* an optional name;
* a class value

Each driver is declared with:
* a major value
* an optional name, descriptive only.

To use a device, its file descriptor (fd) must be looked for. There are
two ways of getting it:
* k_fd(...), if both major and minor are known
* k_fd_byname(...), if name is known.

The class value for a device is a way of shortly declare its kind :
every device having the same class is supposed to handle the same
features, even if they have different major (i.e. different driver).

The pairing device/driver is done inside the k_fd or k_fd_byname
functions: the device is assigned a driver having the same major; if
more than driver has the same major, the behavior is undefined, so it
should not be done.

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>